+++ index.html
@@ -249,7 +249,7 @@
   </section>
 </template>
 
-<script src="./js/main.js"></script>
+<script type="module" src="./js/main.js"></script>
 <script src="./js/functions.js"></script>
 </body>
 </html>
