+++ js/constants.js
@@ -1,0 +1,3 @@
+const PHOTOS_COUNT = 25;
+
+export {PHOTOS_COUNT};
